Fixed round off problem with character position

+++ src/gui/ImagePanel.java
@@ -18,10 +18,10 @@ import java.util.ArrayList;
  */
 public class ImagePanel extends JPanel {
     private BufferedImage image;
-    private int xBlockSize, yBlockSize;
+    private double xBlockSize, yBlockSize;
     private List<Entity> entityList;
 
-    public ImagePanel(BufferedImage img, int xBlkSize, int yBlkSize) {
+    public ImagePanel(BufferedImage img, double xBlkSize, double yBlkSize) {
         image = img;
         xBlockSize = xBlkSize;
         yBlockSize = yBlkSize;
@@ -32,7 +32,7 @@ public class ImagePanel extends JPanel {
 
     // Resizes sprite to block size and adds entity to list
     public void addEntity(Entity e) {
-        e.sprite = Helper.resizeImage(e.sprite, xBlockSize, yBlockSize);
+        e.sprite = Helper.resizeImage(e.sprite, (int)xBlockSize, (int)yBlockSize);
         entityList.add(e);
     }
 
@@ -40,7 +40,7 @@ public class ImagePanel extends JPanel {
     public void paint(Graphics g) {
         g.drawImage(image, 0, 0, null);
         for(Entity e : entityList) {
-            g.drawImage(e.sprite, e.getX() * xBlockSize, e.getY() * yBlockSize, null);
+            g.drawImage(e.sprite, (int)(e.getX() * xBlockSize), (int)(e.getY() * yBlockSize), null);
         }
     }
 }
